refactor(web): migrate Audit, Health, and Insights pages to app dir
Moved `AuditPage.tsx`, `HealthPage.tsx`, and `InsightsPage.tsx` from `web/src/pages/` to `web/src/app/` according to Next.js App Router structural conventions. Updated all relevant imports in `main.tsx` and test files.
