+++ web/src/app/insights/page.tsx
