+++ web/src/main.tsx
@@ -18,9 +18,9 @@ import AgentEditPage from '@/pages/AgentEditPage';
 import AgentNewPage from '@/pages/AgentNewPage';
 import CirclesPage from '@/pages/CirclesPage';
 import CircleDetailPage from '@/pages/CircleDetailPage';
-import InsightsPage from '@/pages/InsightsPage';
-import AuditPage from '@/pages/AuditPage';
-import HealthPage from '@/pages/HealthPage';
+import InsightsPage from '@/app/insights/page';
+import AuditPage from '@/app/audit/page';
+import HealthPage from '@/app/health/page';
 import SchedulesPage from '@/pages/SchedulesPage';
 import SettingsPage from '@/pages/SettingsPage';
 import ToolsPage from '@/pages/ToolsPage';
